Add Router.dispatch
This function synchronizes URL dispatch to a group of routers, waiting
until all have run their transition hooks before proceeding with the
transition. It's also responsible for handling cancelled transitions
and knows how to handle redirects and aborts.

In the case of an abort, the URL reverts to whatever it was before the
last change. Routers remember their current state, so this is essentially
a no-op as far as they're concerned.

Fixes #22