Change the way Match/Miss integrate
The only guarantee from React regarding the lifecycle of ancestors and
descendants is that cDM of descendants will be called before cDM of
ancestors. Also, in the future (React Fibers), cWM may get called but
the component may never get mounted, and there won’t be a lifecycle hook
to know the component’s rendering was cancelled (facebook/react#7671).

So, instead, we add matches in cDM instead of cWM, and then subscribe
to the MatchProvider. When the ancestor MatchProvider mounts, it
notifies the misses so they know whether or not they should render.

This won’t work on the server, but that’s next!