prepare link.isActive for a possible refactor
In the future we may want to use LocationUtils.createLocation to pass
to props.isActive, instead of the simple one in the file right now.

This should let us do it w/o breaking the API … in theory anyway.

I didn’t make that change yet because I don’t want to add another
thing onto context.router, namely createLocation. I think in normal
usage it shouldn’t matter because both the location passed to Link
and the isActive check live in user-land, so don’t give us a
location.search and expect it to be parsed in isActive, just do your
logic on the information you already gave us.